Update capistrano and rvm-capistrano so people using rbenv can deploy
The version of rvm-capistrano we're using doesn't support rbenv.
Support has since been added, so this commit updates us to use that
version.

To update rvm-capistrano we also need to update capistrano. There's a
todo here that updating capistrano caused issues with backup up the
assets manifest file to the release directory.

The asset manifest file used in production is kept in the shared release
folder. Capistrano also makes a copy of this file and puts it in the
directory for the specific release. I read this TODO as relating to this
process, and I couldn't find any other record of it.

After updating capistrano to a newer version of capistrano 2, deploys
are successfully making a copy of this file.

When I run `cap deploy` to my local vagrant box, it runs the copy
successfully:

```
* executing "cp -- /srv/www/shared/assets/.sprockets-manifest-ae51f54419f8b432dd81d41743475379.json /srv/www/releases/20170227041115/assets_manifest.json"
  servers: ["localhost"]
  [localhost:2222] executing command
  command finished in 118ms
```

And a `cksum` of the two files shows they are the same:

```
vagrant@vagrant-ubuntu-trusty-64:~$ cksum /srv/www/shared/assets/.sprockets-manifest-ae51f54419f8b432dd81d41743475379.json /srv/www/releases/20170227024042/assets_manifest.json
4133854102 31161 /srv/www/shared/assets/.sprockets-manifest-ae51f54419f8b432dd81d41743475379.json
4133854102 31161 /srv/www/releases/20170227024042/assets_manifest.json
```

That appears to be resolved to me so I've updated capistrano and removed
the todo.