Allow the deploy user to start and stop the cuttlefish service
Currently the deploy user can only run foreman export and foreman
restart. This means when we changed the recommended startup command in
the readme (be52182fbbe9f2de8b342d180ac9271a73c97a06), we introduced a
regression. The deploy user couldn't actually run this command.

Rather than going back to using the restart command that errors on first
run (because there's no service running to stop), we've decided to allow
the deploy user to start and stop the service.