Use the ansible cron module and set a specific daily time
The cron module uses crontab, rather than placing a file in the
cron.daily folder. This allows us to set a specific time for this
job, which we'd rather not run when other heavy processes are running in
case it crashes.

I've set this to run at 1am, which is an hour after the network and
CPU intensive backup job.