Test the expected behaviour of Archiving.copy_to_s3
This uses the climate_control gem to add the necessary ENV config around
this test, which means devs don't need to do anything special.

It then uses VCR to mock a real call to the AWS api. I used my
credentials and then scrubbed anything I thought looked vaugely useful
to hackers from the cassette file.

The test looks for a valid Fog Storage File response, which means the
file got there successfully.

This also adds a fixture valid archive file to the repo, which I
generated using the exiting factories in this test file. So in other
words, that's just the data that's already in this test, not anything
private.