Build click_event
Currently this variable doesn't get called so the click event is never
created. This change fires the create event so it is.

This variable isn't used in the test, but I assume that the data is
being set to test that the Archive.serialise and Archive.deserialise
methods handle all this data. So I'm creating it so it actually does
that now.