Remove pull down for selecting metadata key. It was too slow
It looks like too much work (and a bit ugly) to make the database query 
underlying the pulldown not be slow.

So, decided to remove it. We can still filter by key. We just need to 
enter the url by hand or select it from the detail view of an email 
you've already found.